+++ src/javascript/9000-main.js
@@ -1,14 +1,17 @@
-s.version.set(0,5,3);
+/* Builds the node structure for the game. Nodes, Actions, and Outcomes
+should never be created dynamically. They should all be constructed
+here, prior to the start of play. Use an Action's check function to hide
+or show possible actions, depending on conditions. Use an Actions's
+choose function to select between different possible outcomes, depending
+on conditions.
+*/
+
+s.version.set(0,5,4);
 s.savesMgr.setBlock(true); // block loading of saves from old versions
 s.savesMgr.setBkMarks(8); // allow 8 bookmarks instead of 1
 st.page.setContinuous(true); // make embedding passages the default
 
-s.root = new s.Node('intro')
-
 // build menu
-s.menu.onBegin(function() {
-    st.page.load(s.root);
-});
 s.menu.addInfoNode('info', 'help');
 s.menu.addInfoNode('bio', 'bio');
 
@@ -18,10 +21,10 @@ s.debCon.setCheat([
 ]);
 
 // make initial links
-s.addLink('intro', 'continue', null, 'RNG warning');
+s.addLink('Root', 'continue', null, 'RNG warning');
 s.addLink(
-    'RNG warning', 'got it', null, 'who you are', null, false // don't embed
-);
+    'RNG warning', 'got it', null, 'who you are', null
+).setEmbed(false); // don't embed
 s.addLink('who you are', 'take a look around', null, 'bedroom');
 
 // bedroom actions
@@ -39,9 +42,8 @@ s.addLink('who you are', 'take a look around', null, 'bedroom');
         function() {
             // check
             return (v.containedIn.get('knife') === 'bedroom');
-        },
-        true, true // embed with scene break
-    );
+        }
+    ).setEmbed(true, true); // embed without scene break
 
     // search the body
     s.setSubCount('Cops bust in', 2);